+++ packages/core/styles/src/packages/atomic/tooltip/nimbus-tooltip.css.ts
@@ -8,7 +8,7 @@ export const container = style({
 
 export const content = style({
   width: "fit-content",
-  backgroundColor: varsThemeBase.colors.neutral.textLow,
+  backgroundColor: varsThemeBase.colors.neutral.textHigh,
   padding: varsThemeBase.spacing[1],
   borderRadius: varsThemeBase.shape.border.radius[1],
   transition: `opacity ${varsThemeBase.motion.speed.fast} ease`,