+++ packages/react/src/composite/Modal/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Modal component allows us to call the user's attention to a floating box that can have text, actions or forms to perform tasks by changing the focus from the background. It is an intrusive component as it interrupts the user's operation to present a message or content.
 
+## 2023-12-22 `1.5.1`
+
+#### 🎉 New features
+
+- Changed default title size to `h4` on `ModalHeader` subcomponent. ([#215](https://github.com/TiendaNube/nimbus-design-system/pull/215) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2023-10-13 `1.5.0`
 
 #### 📚 3rd party library updates